# Conflicts:
#	src/main/java/controleur/Controleur.java
#	target/classes/controleur/Controleur$1.class
#	target/classes/controleur/Controleur.class

+++ src/main/java/modele/Protagoniste.java
@@ -1,6 +1,8 @@
 package modele;
 
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 
 public class Protagoniste {
@@ -9,32 +11,33 @@ public class Protagoniste {
     private String main1;
     private String armure;
 
-    private IntegerProperty xProperty;
+    private DoubleProperty xProperty;
 
-    private IntegerProperty yProperty;
+    private DoubleProperty yProperty;
 
     public Protagoniste(){
         this.vie = 20;
         this.main1 = null;
         this.armure = null;
-        xProperty = new SimpleIntegerProperty(0);
-        yProperty = new SimpleIntegerProperty(0);
+        xProperty = new SimpleDoubleProperty(0);
+        yProperty = new SimpleDoubleProperty(0);
     }
 
-    public final IntegerProperty getXProperty() {
+    public final DoubleProperty getXProperty() {
         return xProperty;
     }
 
-    public  final IntegerProperty getYProperty() {
+    public  final DoubleProperty getYProperty() {
         return yProperty;
     }
 
-    public final void setXProperty(int nb) {
+    public final void setXProperty(double nb) {
         xProperty.setValue(nb);
     }
 
-    public final void setYProperty(int nb) {
+    public final void setYProperty(double nb) {
         yProperty.setValue(nb);
     }
+
 
 }
